Added App Layout using the Plotly Dash library
The app is composed of a series of drop-down menus, images, buttons, and text elements that allow the user to select various attributes of a UFC fight, such as the weight class, the number of rounds, the type of fight, and the fighters. The "Predict" button is expected to trigger a prediction of the outcome of the fight based on these attributes, and the app is expected to display the prediction as a probability for each fighter.