Create HTML_Parsing.py
Using the requests library to retrieve the source code of a given URL and the BeautifulSoup library to parse the HTML content of the page. The function make_soup takes a URL as an argument and returns a BeautifulSoup object, which can be used to extract information from the page.